requirement.txt
requests==2.26.0
beautifulsoup4==4.10.0
newspaper3k==0.2.8
